+++ packages/halo/src/vars/index.js
@@ -1,9 +1,0 @@
-export const gapSizes = {
-  default: 'default',
-  small: 'small'
-}
-
-export const shapes = {
-  default: 'default',
-  pill: 'pill'
-}