+++ packages/starrating/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-starrating@7.2.0...@pluralsight/ps-design-system-starrating@7.2.1) (2020-10-13)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-starrating
+
+
+
+
+
+# [7.2.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-starrating@7.1.18...@pluralsight/ps-design-system-starrating@7.2.0) (2020-10-13)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-starrating
+
+
+
+
+
+## [7.1.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-starrating@7.1.17...@pluralsight/ps-design-system-starrating@7.1.18) (2020-10-13)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-starrating
+
+
+
+
+
 ## [7.1.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-starrating@7.1.16...@pluralsight/ps-design-system-starrating@7.1.17) (2020-10-08)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-starrating